implemented simple ANNs and test main with GAs
the ANN is pretty barebones but should be enough for our purposes.

The __main__ of the neural_GA.py will run a short genetic algorithm with ANNs that tries to optimize autoencoding a bitstring of 8 bits (only one of them is 1, the rest is 0) using a latent space of 3 neurons. This is a task that is usually very doable with backpropagation, but seems to be a little more challenging with GAs.